zsh/profile: Drop all systemctl queries
Adding the is-system-active query caused the graphical environment to
not be started because man-db was still running in the background and
systemd considered the system to be still booting. Removing all
systemctl commands will stop that from happening, and will also fix
errors on other environments that don't even have GUIs like Termux
and WSL. Also, I don't really mind that these commands will run even
if I'm in multi-user.target instead of graphical.target.